requirements.txt
fastapi
uvicorn
python-multipart
python-docx
spacy
pdfplumber
python-docx
striprtf
odfpy
pydantic